Move MethodData and IndexerData to their own files.
